Move parameditorview to own file
